moving evaluate_dataset to tfodeval
